Created Products and Catalog Classes
Created products and catalog CRUD classes. Added additional method to dbcon class to manage queries from products and catalog classes. The method returns an array back of all of the data returned from the query. Not sure how this will work for update, delete, and create requests, but I can create a new method for those as necessary in the dbcon class.