Created Product and Catalog Handler classes
Created a product and catalog handler class to handle ajax requests to the server to update the DB. Drop down boxes work, create, update, and delete works for catalogs and products. Still working on how to populate the product textfields with json that we are receiving from the server. We are getting the encoded json back but Im having trouble parsing it and populating the fields. Need to further troubleshoot.