Updates to Catalog
Updated catalog to display data from database. Need to create an object of all of the selected items on the page and pass it over to the cart  page to display all of the sale items the customer selected. Once we have that data displayed they can click the submit button to process a transaction which will create a transaction record and add that transaction id to all of the sale items in the transaction.

Made small changes to admin panel removing all of the extra document.ready statements that were unnecessary.