Updated several classes
Got the sale item to create once the user selects items on the catalog page. Currently no way to tie the saleitem to a specific user. Once we discuss and figure out how we would like to tie the saleitem to a user we can incorporate that with the sale item and have items without a transaction id come up on the users cart. We discussed using the session to grab the users data which is what we will need to do for the cart as well as the catalog page. I can pass the user data to the sale item and add another column in the sale item table to track the user.